Fix for create_or_modify_pyi read_text Windows Issue in component_meta.py (#9242)
* Fix for create_or_modify_pyi read_text in component_meta.py

In some environments the `read_text()` call in `create_or_modify_pyi` will not use expected encoding when reading the `source_file`. In Unix environments (and presumably MacOS though I have not tested), this will default to use the 'utf-8' encoding, which is what the `source_file` are encoded in. In Windows, it seems to default to a different encoding when 'utf-8' should be selected.

* add changeset

* format

---------

Co-authored-by: gradio-pr-bot <gradio-pr-bot@users.noreply.github.com>
Co-authored-by: Abubakar Abid <abubakar@huggingface.co>